Adds detection for Telma, Quipus brands and improves detection for ClearPHONE, Cloudfone, Gol Mobile, Hammer, iSWAG, Konka, Mintt, SPC, T-Mobile, Tesla, Vivax, Zebra brands (matomo-org#7263)
Adds detection for Cloudfone Thrill Power
Adds detection for Telma F1 Max 4G
Adds detection for Quipus Tinku
Adds detection for T-Mobile REVVL 6 5G, REVVL 6 Pro 5G
Adds detection for Gol Mobile Team 7.0 3G
Adds detection for Vivax TPC-806 3G
Adds detection for Zebra TC15
Adds detection for Zebra TC77
Adds detection for Hammer Energy 2 (2022)
Adds detection for SPC Gravity 2 Max
Adds detection for Konka Phoenix, Phoenix Pro
Adds detection for ClearPHONE 220
Adds detection for iSWAG Neo
Adds detection for Mintt CoolMintt Blaze 2
Adds detection for iSWAG Activ
Improves detection for Tesla

Detect devices for existing brands (matomo-org#7261)
feat(device) detect brand Realme: V20 (RMX3610)
feat(device) detect brand UMIDIGI: Bison GT2 (RP03), Bison 2 (RP04)
feat(device) detect brand OPPO: Find X5 Pro (PFFM20)
feat(device) detect brand OnePlus: Ace Pro (PGP110)
feat(device) detect brand POCO: F4 GT
feat(device) detect brand Sony: Xperia Pro (XQ-AQ62)
feat(device) detect new brand BOOX: Nova Air C
feat(device) detect new brand Rizzen: NovaTab R10
feat(device) detect brand Xiaomi: Redmi K40 Gaming, Redmi 9T
feat(device) detect brand Samsung: Galaxy A13 5G (SM-A136B), Galaxy F13 (SM-E135F), Galaxy S21+ 5G (SM-G996P)
feat(device) detect brand Realme: V25 (RMX3475), C35 (RMX3513)
feat(device) detect brand OPPO: Reno 8 Pro (CPH2357), Find N
feat(device) detect brand Mobicel: Epic Pro 1
feat(device) detect brand Huawei: Mate XS 2 (PAL-AL00, PAL-LX9)
feat(device) detect brand IKU Mobile: A22
feat(device) detect brand Xiaomi: Black Shark 4S, Black Shark 4S Pro, Black Shark 4, Black Shark 4 Pro
feat(device) detect brand Gplus: P10+
feat(device) detect brand IKU Mobile: X2
feat(device) detect brand OKSI: GT1001
feat(device) detect brand iHunt: Like 12 Pro, Titan X 5G, S22 Ultra, S22 Plus
feat(device) detect brand ONYX BOOX: Nova Air 2, Nova Air 2 Plus
feat(device) detect brand Samsung: Galaxy A12 (SM-A125A, SM-A125T)
feat(device) detect brand BQ: Clever (BQ-5765L), Wide (BQru-6868L)
feat(device) detect brand BMAX: MaxPad I10 Pro
feat(device) detect brand DEXP: K61, D11, D21
feat(device) detect brand Technopc: UltraPad UP10SH36LAG
feat(device) detect brand Teclast: T40 Pro
feat(device) rename model for brand BQ: Fast => Wide

+++ tests/tools/update-fixtures.js
@@ -1,6 +1,15 @@
 const spawn = require('node:child_process').spawn;
 const fs = require('node:fs');
+
 const TMP_DIR = __dirname + '/../output/_git';
+
+const errorLog = (signal) => {
+  console.log(signal.toString());
+};
+
+const log = (signal) => {
+  console.log(signal.toString());
+};
 
 const gitDownload = () => {
   return new Promise((resolve, reject) => {
@@ -9,9 +18,12 @@ const gitDownload = () => {
         'clone', 'git@github.com:matomo-org/device-detector.git', TMP_DIR],
     ];
     const git = spawn.apply(spawn, cmd);
-    git.stderr.on('data', console.error);
+    git.stderr.on('data', errorLog);
+    git.stdin.on('data', log);
     git.on('close', function(err, signal) {
-      if (err) return reject(err);
+      if (err){
+        return reject(err);
+      }
       return resolve(signal);
     });
   });
@@ -156,16 +168,48 @@ const syncShortBrowserFamilies = async () => {
   )
 };
 
+const syncShortOsFamilies = async () => {
+  console.log('Sync os families');
+  const sourceFile = TMP_DIR + '/Parser/OperatingSystem.php';
+  const targetFile = __dirname + '/../../parser/os/os_families.js';
+  const content = fs.readFileSync(sourceFile).toString();
+  const match = / static \$osFamilies = \[([^;]+)\];/mg.exec(content);
+  
+  let newContent = match[1]
+  .replace(/\s+=>\s+/gm, ": ")
+  .replace(/^\s+/gm, "  ")
+  
+  fs.writeFileSync(targetFile,
+    `// prettier-ignore\nmodule.exports = {\n${newContent}\n};\n`
+  )
+};
+
+const syncShortOs = async () => {
+  console.log('Sync os');
+  const sourceFile = TMP_DIR + '/Parser/OperatingSystem.php';
+  const targetFile = __dirname + '/../../parser/os/os_systems.js';
+  const content = fs.readFileSync(sourceFile).toString();
+  const match = / static \$operatingSystems = \[([^;]+)\];/mg.exec(content);
+  
+  let newContent = match[1]
+  .replace(/\s+=>\s+/gm, ": ")
+  .replace(/^\s+/gm, "  ")
+  
+  fs.writeFileSync(targetFile,
+    `// prettier-ignore\nmodule.exports = {\n${newContent}\n};\n`
+  )
+};
+
 (async function() {
-  /*if (fs.existsSync(TMP_DIR)) {
-      console.log('clear dir output/_git');
-      fs.rmSync(TMP_DIR, {recursive: true, force: true});
+  if (fs.existsSync(TMP_DIR)) {
+    console.log('clear dir output/_git');
+    fs.rmSync(TMP_DIR, {recursive: true, force: true});
   }
   console.log('create dir output/_git');
   fs.mkdirSync(TMP_DIR);
   console.log('download git repository');
-  await gitDownload();*/
   
+  await gitDownload();
   await syncTestOsFixtures();
   await syncTestDeviceFixtures();
   await syncTestParserFixtures();
@@ -176,6 +220,8 @@ const syncShortBrowserFamilies = async () => {
   await syncShortBrowserBrands();
   await syncShortMobileBrowser();
   await syncShortBrowserFamilies();
+  await syncShortOsFamilies();
+  await syncShortOs();
   
 })();
